Move files into src directory
