+++ test/test-state.js
@@ -3,11 +3,11 @@
 const fs = require('fs')
 const chai = require('chai')
 const chaiHttp = require('chai-http')
-const constants = require('../constants.js')
+const constants = require('../src/constants.js')
 const accounts = require('./mock-accounts.js').accounts
-const web3 = require('../eth.js')
+const web3 = require('../src/eth.js')
 const BN = web3.utils.BN
-const State = require('../state.js')
+const State = require('../src/state.js')
 const levelup = require('levelup')
 const leveldown = require('leveldown')
 const encoder = require('plasma-utils').encoder
